+++ libs/rxjs/utility/src/lib/temperature.spec.ts
@@ -38,11 +38,11 @@ describe('temperature', () => {
         });
         const subs = '^---------!';
         const expected = m.cold('-a-b-c-d-e|', {
-          a: 273.2,
-          b: 373.2,
-          c: 310.7,
-          d: 231.2,
-          e: 100273.2,
+          a: 273.15,
+          b: 373.15,
+          c: 310.65,
+          d: 231.15,
+          e: 100273.15,
         });
         m.expect(input.pipe(temperature(Temperatures.CELSIUS, Temperatures.KELVIN))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -61,11 +61,11 @@ describe('temperature', () => {
         });
         const subs = '^---------!';
         const expected = m.cold('-a-b-c-d-e|', {
-          a: 491.7,
-          b: 671.7,
-          c: 559.2,
-          d: 416.1,
-          e: 180491.7,
+          a: 491.67,
+          b: 671.67,
+          c: 559.17,
+          d: 416.07,
+          e: 180491.67,
         });
         m.expect(input.pipe(temperature(Temperatures.CELSIUS, Temperatures.RANKINE))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -73,7 +73,7 @@ describe('temperature', () => {
     );
 
     it(
-      'should return to 2 decimal places',
+      'should return to 3 decimal places',
       marbles((m) => {
         const input = m.hot('-a-b-c-d-e|', {
           a: 0,
@@ -85,12 +85,12 @@ describe('temperature', () => {
         const subs = '^---------!';
         const expected = m.cold('-a-b-c-d-e|', {
           a: 0,
-          b: 100.39,
-          c: 37.54,
+          b: 100.393,
+          c: 37.536,
           d: -42.32,
           e: 100000.3,
         });
-        m.expect(input.pipe(temperature(Temperatures.CELSIUS, Temperatures.CELSIUS, 2))).toBeObservable(expected);
+        m.expect(input.pipe(temperature(Temperatures.CELSIUS, Temperatures.CELSIUS, 3))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -132,11 +132,11 @@ describe('temperature', () => {
         });
         const subs = '^---------!';
         const expected = m.cold('-a-b-c-d-e|', {
-          a: 255.4,
-          b: 310.9,
-          c: 276.2,
-          d: 232,
-          e: 55810.9,
+          a: 255.37,
+          b: 310.93,
+          c: 276.21,
+          d: 232.04,
+          e: 55810.93,
         });
         m.expect(input.pipe(temperature(Temperatures.FAHRENHEIT, Temperatures.KELVIN))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -155,11 +155,11 @@ describe('temperature', () => {
         });
         const subs = '^---------!';
         const expected = m.cold('-a-b-c-d-e|', {
-          a: 459.7,
-          b: 559.7,
-          c: 497.2,
-          d: 417.7,
-          e: 100459.7,
+          a: 459.67,
+          b: 559.67,
+          c: 497.17,
+          d: 417.67,
+          e: 100459.67,
         });
         m.expect(input.pipe(temperature(Temperatures.FAHRENHEIT, Temperatures.RANKINE))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -167,7 +167,7 @@ describe('temperature', () => {
     );
 
     it(
-      'should return to 2 decimal places',
+      'should return to 3 decimal places',
       marbles((m) => {
         const input = m.hot('-a-b-c-d-e|', {
           a: 0,
@@ -179,12 +179,12 @@ describe('temperature', () => {
         const subs = '^---------!';
         const expected = m.cold('-a-b-c-d-e|', {
           a: 0,
-          b: 100.39,
-          c: 37.54,
+          b: 100.393,
+          c: 37.536,
           d: -42.32,
           e: 100000.3,
         });
-        m.expect(input.pipe(temperature(Temperatures.FAHRENHEIT, Temperatures.FAHRENHEIT, 2))).toBeObservable(expected);
+        m.expect(input.pipe(temperature(Temperatures.FAHRENHEIT, Temperatures.FAHRENHEIT, 3))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -203,11 +203,11 @@ describe('temperature', () => {
         });
         const subs = '^---------!';
         const expected = m.cold('-a-b-c-d-e|', {
-          a: -241.1,
-          b: -61.1,
-          c: -173.6,
-          d: -316.7,
-          e: 179758.9,
+          a: -241.15,
+          b: -61.15,
+          c: -173.65,
+          d: -316.75,
+          e: 179758.85,
         });
         m.expect(input.pipe(temperature(Temperatures.KELVIN, Temperatures.CELSIUS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -226,11 +226,11 @@ describe('temperature', () => {
         });
         const subs = '^---------!';
         const expected = m.cold('-a-b-c-d-e|', {
-          a: -459.7,
-          b: -279.7,
-          c: -392.2,
-          d: -535.3,
-          e: 179540.3,
+          a: -459.67,
+          b: -279.67,
+          c: -392.17,
+          d: -535.27,
+          e: 179540.33,
         });
         m.expect(input.pipe(temperature(Temperatures.KELVIN, Temperatures.FAHRENHEIT))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -261,7 +261,7 @@ describe('temperature', () => {
     );
 
     it(
-      'should return to 2 decimal places',
+      'should return to 3 decimal places',
       marbles((m) => {
         const input = m.hot('-a-b-c-d-e|', {
           a: 0,
@@ -273,12 +273,12 @@ describe('temperature', () => {
         const subs = '^---------!';
         const expected = m.cold('-a-b-c-d-e|', {
           a: 0,
-          b: 100.39,
-          c: 37.54,
+          b: 100.393,
+          c: 37.536,
           d: -42.32,
           e: 100000.3,
         });
-        m.expect(input.pipe(temperature(Temperatures.KELVIN, Temperatures.KELVIN, 2))).toBeObservable(expected);
+        m.expect(input.pipe(temperature(Temperatures.KELVIN, Temperatures.KELVIN, 3))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -297,11 +297,11 @@ describe('temperature', () => {
         });
         const subs = '^---------!';
         const expected = m.cold('-a-b-c-d-e|', {
-          a: -827.4,
-          b: -503.4,
-          c: -705.9,
-          d: -963.5,
-          e: 323172.6,
+          a: -827.41,
+          b: -503.41,
+          c: -705.91,
+          d: -963.49,
+          e: 323172.59,
         });
         m.expect(input.pipe(temperature(Temperatures.RANKINE, Temperatures.CELSIUS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -320,11 +320,11 @@ describe('temperature', () => {
         });
         const subs = '^---------!';
         const expected = m.cold('-a-b-c-d-e|', {
-          a: -491.7,
-          b: -391.7,
-          c: -454.2,
-          d: -533.7,
-          e: 99508.3,
+          a: -491.67,
+          b: -391.67,
+          c: -454.17,
+          d: -533.67,
+          e: 99508.33,
         });
         m.expect(input.pipe(temperature(Temperatures.RANKINE, Temperatures.FAHRENHEIT))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -344,10 +344,10 @@ describe('temperature', () => {
         const subs = '^---------!';
         const expected = m.cold('-a-b-c-d-e|', {
           a: 0,
-          b: 55.6,
-          c: 20.8,
-          d: -23.3,
-          e: 55555.6,
+          b: 55.56,
+          c: 20.83,
+          d: -23.33,
+          e: 55555.56,
         });
         m.expect(input.pipe(temperature(Temperatures.RANKINE, Temperatures.KELVIN))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -355,7 +355,7 @@ describe('temperature', () => {
     );
 
     it(
-      'should return to 2 decimal places',
+      'should return to 3 decimal places',
       marbles((m) => {
         const input = m.hot('-a-b-c-d-e|', {
           a: 0,
@@ -367,12 +367,12 @@ describe('temperature', () => {
         const subs = '^---------!';
         const expected = m.cold('-a-b-c-d-e|', {
           a: 0,
-          b: 100.39,
-          c: 37.54,
+          b: 100.393,
+          c: 37.536,
           d: -42.32,
           e: 100000.3,
         });
-        m.expect(input.pipe(temperature(Temperatures.RANKINE, Temperatures.RANKINE, 2))).toBeObservable(expected);
+        m.expect(input.pipe(temperature(Temperatures.RANKINE, Temperatures.RANKINE, 3))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );