+++ libs/rxjs/utility/src/lib/weight.spec.ts
@@ -5,7 +5,7 @@ describe('weight', () => {
   // Grams
   describe('grams', () => {
     it(
-      'should return same value with precision 2',
+      'should return same value with precision 3',
       marbles((m) => {
         const input = m.hot('-a-b-c-|', {
           a: 0,
@@ -15,10 +15,10 @@ describe('weight', () => {
         const subs = '^------!';
         const expected = m.cold('-a-b-c-|', {
           a: 0,
-          b: 100.26,
-          c: 1000.25,
+          b: 100.257,
+          c: 1000.254,
         });
-        m.expect(input.pipe(weight(Weights.GRAMS, Weights.GRAMS, 2))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.GRAMS, Weights.GRAMS, 3))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -53,7 +53,7 @@ describe('weight', () => {
         const expected = m.cold('-a-b-c-|', {
           a: 0,
           b: 0.22,
-          c: 2.205,
+          c: 2.2,
         });
         m.expect(input.pipe(weight(Weights.GRAMS, Weights.POUNDS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -71,8 +71,8 @@ describe('weight', () => {
         const subs = '^------!';
         const expected = m.cold('-a-b-c-|', {
           a: 0,
-          b: 3.527,
-          c: 35.274,
+          b: 3.53,
+          c: 35.27,
         });
         m.expect(input.pipe(weight(Weights.GRAMS, Weights.OUNCES))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -90,8 +90,8 @@ describe('weight', () => {
         const subs = '^------!';
         const expected = m.cold('-a-b-c-|', {
           a: 0,
-          b: 0.016,
-          c: 0.157,
+          b: 0.02,
+          c: 0.16,
         });
         m.expect(input.pipe(weight(Weights.GRAMS, Weights.STONE))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -102,7 +102,7 @@ describe('weight', () => {
   // Kilograms
   describe('kilograms', () => {
     it(
-      'should return same value with precision 2',
+      'should return same value with precision 3',
       marbles((m) => {
         const input = m.hot('-a-b-c-|', {
           a: 0,
@@ -112,10 +112,10 @@ describe('weight', () => {
         const subs = '^------!';
         const expected = m.cold('-a-b-c-|', {
           a: 0,
-          b: 100.26,
-          c: 1000.25,
+          b: 100.257,
+          c: 1000.254,
         });
-        m.expect(input.pipe(weight(Weights.KILOGRAMS, Weights.KILOGRAMS, 2))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.KILOGRAMS, Weights.KILOGRAMS, 3))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -199,7 +199,7 @@ describe('weight', () => {
   // Pounds
   describe('Pounds', () => {
     it(
-      'should return same value with precision 2',
+      'should return same value with precision 3',
       marbles((m) => {
         const input = m.hot('-a-b-c-|', {
           a: 0,
@@ -209,10 +209,10 @@ describe('weight', () => {
         const subs = '^------!';
         const expected = m.cold('-a-b-c-|', {
           a: 0,
-          b: 100.26,
-          c: 1000.25,
+          b: 100.257,
+          c: 1000.254,
         });
-        m.expect(input.pipe(weight(Weights.POUNDS, Weights.POUNDS, 2))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.POUNDS, Weights.POUNDS, 3))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -227,8 +227,8 @@ describe('weight', () => {
         const subs = '^------!';
         const expected = m.cold('-a-b-c-|', {
           a: 0,
-          b: 45359.702,
-          c: 453597.024,
+          b: 45359.7,
+          c: 453597.02,
         });
         m.expect(input.pipe(weight(Weights.POUNDS, Weights.GRAMS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -247,7 +247,7 @@ describe('weight', () => {
         const expected = m.cold('-a-b-c-|', {
           a: 0,
           b: 45.36,
-          c: 453.597,
+          c: 453.6,
         });
         m.expect(input.pipe(weight(Weights.POUNDS, Weights.KILOGRAMS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -284,8 +284,8 @@ describe('weight', () => {
         const subs = '^------!';
         const expected = m.cold('-a-b-c-|', {
           a: 0,
-          b: 7.143,
-          c: 71.429,
+          b: 7.14,
+          c: 71.43,
         });
         m.expect(input.pipe(weight(Weights.POUNDS, Weights.STONE))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -296,7 +296,7 @@ describe('weight', () => {
   // Ounces
   describe('Ounces', () => {
     it(
-      'should return same value with precision 2',
+      'should return same value with precision 3',
       marbles((m) => {
         const input = m.hot('-a-b-c-|', {
           a: 0,
@@ -306,10 +306,10 @@ describe('weight', () => {
         const subs = '^------!';
         const expected = m.cold('-a-b-c-|', {
           a: 0,
-          b: 100.26,
-          c: 1000.25,
+          b: 100.257,
+          c: 1000.254,
         });
-        m.expect(input.pipe(weight(Weights.OUNCES, Weights.OUNCES, 2))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.OUNCES, Weights.OUNCES, 3))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -324,8 +324,8 @@ describe('weight', () => {
         const subs = '^------!';
         const expected = m.cold('-a-b-c-|', {
           a: 0,
-          b: 2834.949,
-          c: 28349.493,
+          b: 2834.95,
+          c: 28349.49,
         });
         m.expect(input.pipe(weight(Weights.OUNCES, Weights.GRAMS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -343,8 +343,8 @@ describe('weight', () => {
         const subs = '^------!';
         const expected = m.cold('-a-b-c-|', {
           a: 0,
-          b: 2.835,
-          c: 28.349,
+          b: 2.83,
+          c: 28.35,
         });
         m.expect(input.pipe(weight(Weights.OUNCES, Weights.KILOGRAMS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -381,8 +381,8 @@ describe('weight', () => {
         const subs = '^------!';
         const expected = m.cold('-a-b-c-|', {
           a: 0,
-          b: 0.446,
-          c: 4.464,
+          b: 0.45,
+          c: 4.46,
         });
         m.expect(input.pipe(weight(Weights.OUNCES, Weights.STONE))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -393,7 +393,7 @@ describe('weight', () => {
   // Stones
   describe('Stones', () => {
     it(
-      'should return same value with precision 2',
+      'should return same value with precision 3',
       marbles((m) => {
         const input = m.hot('-a-b-c-|', {
           a: 0,
@@ -403,10 +403,10 @@ describe('weight', () => {
         const subs = '^------!';
         const expected = m.cold('-a-b-c-|', {
           a: 0,
-          b: 100.26,
-          c: 1000.25,
+          b: 100.257,
+          c: 1000.254,
         });
-        m.expect(input.pipe(weight(Weights.STONE, Weights.STONE, 2))).toBeObservable(expected);
+        m.expect(input.pipe(weight(Weights.STONE, Weights.STONE, 3))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
       }),
     );
@@ -421,8 +421,8 @@ describe('weight', () => {
         const subs = '^------!';
         const expected = m.cold('-a-b-c-|', {
           a: 0,
-          b: 635041.595,
-          c: 6350415.952,
+          b: 635041.6,
+          c: 6350415.95,
         });
         m.expect(input.pipe(weight(Weights.STONE, Weights.GRAMS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);
@@ -440,8 +440,8 @@ describe('weight', () => {
         const subs = '^------!';
         const expected = m.cold('-a-b-c-|', {
           a: 0,
-          b: 635.042,
-          c: 6350.416,
+          b: 635.04,
+          c: 6350.42,
         });
         m.expect(input.pipe(weight(Weights.STONE, Weights.KILOGRAMS))).toBeObservable(expected);
         m.expect(input).toHaveSubscriptions(subs);